Add specs for the Auth endpoint
Add tests for request with correct credentials to test if
the response is a `Hashie::Mash` and that the response
includes `status` and `paymentState`.

Add tests for request with the wrong credentials to test if
the error uses `Afterpay::PaymentRequiredError` and the message
is `Cannot complete payment, expired or invalid token.`.