Add the void endpoint class
Add the `@order_id` as this is needed for the endpoint
to be able to void the correct order.

Add the attr_accessor `amount` to the payment component as it is
required for the void endpoint.

require the void component in the `afterpay.rb` file as it
is needed in order to use the void class.