Add specs for void and amount to payment component
Add tests for when giving the correct `order_id` to
test if it returns a `Hashie::Mash` and if it includes
the `paymentState`.

Add tests for when giving the wrong `order_id` to
test if the error is from `Afterpay::NotFoundError` and
the error message is `Afterpay payment ID not found.`.